Creado componente Buscador y separado funcion 'buscar'
